Refactor geometries, support error checking on reset geometries
