+++ hoomd/mpcd/ConfinedStreamingMethodGPU.h
@@ -43,7 +43,7 @@ class ConfinedStreamingMethodGPU : public mpcd::ConfinedStreamingMethod<Geometry
                                    unsigned int cur_timestep,
                                    unsigned int period,
                                    int phase,
-                                   std::shared_ptr<Geometry> geom)
+                                   const Geometry& geom)
             : mpcd::ConfinedStreamingMethod<Geometry>(sysdata, cur_timestep, period, phase, geom)
             {
             m_tuner.reset(new Autotuner(32, 1024, 32, 5, 100000, "mpcd_stream", this->m_exec_conf));
@@ -77,7 +77,7 @@ void ConfinedStreamingMethodGPU<Geometry>::stream(unsigned int timestep)
     {
     if (!this->shouldStream(timestep)) return;
 
-    // the validation step currently proceeds on the cpu because it does done infrequently
+    // the validation step currently proceeds on the cpu because it is done infrequently.
     // if it becomes a performance concern, it can be ported to the gpu
     if (this->m_validate_geom)
         {
@@ -96,7 +96,7 @@ void ConfinedStreamingMethodGPU<Geometry>::stream(unsigned int timestep)
                                   m_tuner->getParam());
 
     m_tuner->begin();
-    mpcd::gpu::confined_stream<Geometry>(args, *(this->m_geom));
+    mpcd::gpu::confined_stream<Geometry>(args, this->m_geom);
     if (this->m_exec_conf->isCUDAErrorCheckingEnabled()) CHECK_CUDA_ERROR();
     m_tuner->end();
 
@@ -118,7 +118,7 @@ void export_ConfinedStreamingMethodGPU(pybind11::module& m)
     const std::string name = "ConfinedStreamingMethodGPU" + Geometry::getName();
     py::class_<mpcd::ConfinedStreamingMethodGPU<Geometry>, std::shared_ptr<mpcd::ConfinedStreamingMethodGPU<Geometry>>>
         (m, name.c_str(), py::base<mpcd::ConfinedStreamingMethod<Geometry>>())
-        .def(py::init<std::shared_ptr<mpcd::SystemData>, unsigned int, unsigned int, int, std::shared_ptr<Geometry>>());
+        .def(py::init<std::shared_ptr<mpcd::SystemData>, unsigned int, unsigned int, int, const Geometry&>());
     }
 } // end namespace detail
 } // end namespace mpcd
