+++ hoomd/mpcd/test-py/stream_slit_test.py
@@ -40,28 +40,28 @@ class mpcd_stream_slit_test(unittest.TestCase):
         self.assertAlmostEqual(slit.H, 4.)
         self.assertAlmostEqual(slit.V, 0.)
         self.assertEqual(slit.boundary, "no_slip")
-        self.assertAlmostEqual(slit._geometry.H, 4.)
-        self.assertAlmostEqual(slit._geometry.V, 0.)
-        self.assertEqual(slit._geometry.boundary, mpcd._mpcd.boundary.no_slip)
+        self.assertAlmostEqual(slit._cpp.geometry.H, 4.)
+        self.assertAlmostEqual(slit._cpp.geometry.V, 0.)
+        self.assertEqual(slit._cpp.geometry.boundary, mpcd._mpcd.boundary.no_slip)
 
         # change H and also ensure other parameters stay the same
         slit.set_params(H=2.)
         self.assertAlmostEqual(slit.H, 2.)
         self.assertAlmostEqual(slit.V, 0.)
         self.assertEqual(slit.boundary, "no_slip")
-        self.assertAlmostEqual(slit._geometry.H, 2.)
-        self.assertAlmostEqual(slit._geometry.V, 0.)
-        self.assertEqual(slit._geometry.boundary, mpcd._mpcd.boundary.no_slip)
+        self.assertAlmostEqual(slit._cpp.geometry.H, 2.)
+        self.assertAlmostEqual(slit._cpp.geometry.V, 0.)
+        self.assertEqual(slit._cpp.geometry.boundary, mpcd._mpcd.boundary.no_slip)
 
         # change V
         slit.set_params(V=0.1)
         self.assertAlmostEqual(slit.V, 0.1)
-        self.assertAlmostEqual(slit._geometry.V, 0.1)
+        self.assertAlmostEqual(slit._cpp.geometry.V, 0.1)
 
         # change BCs
         slit.set_params(boundary="slip")
         self.assertEqual(slit.boundary, "slip")
-        self.assertEqual(slit._geometry.boundary, mpcd._mpcd.boundary.slip)
+        self.assertEqual(slit._cpp.geometry.boundary, mpcd._mpcd.boundary.slip)
 
     # test for invalid boundary conditions being set
     def test_bad_boundary(self):
@@ -160,12 +160,19 @@ class mpcd_stream_slit_test(unittest.TestCase):
 
     # test that setting the slit size too large raises an error
     def test_validate_box(self):
+        # initial configuration is invalid
         slit = mpcd.stream.slit(H=10.)
         with self.assertRaises(RuntimeError):
             hoomd.run(1)
 
+        # now it should be valid
         slit.set_params(H=2.)
         hoomd.run(1)
+
+        # make sure we can invalidate it again
+        slit.set_params(H=10.)
+        with self.assertRaises(RuntimeError):
+            hoomd.run(1)
 
     # test that particles out of bounds can be caught
     def test_out_of_bounds(self):
